pyproject.toml
[tool.poetry]
name = "pycon-opentelemetry-demo"
version = "0.1.0"
description = ""
authors = ["hsuyuming <abego452@gmail.com>"]
readme = "README.md"
packages = [{include = "pycon_opentelemetry_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
opentelemetry-sdk = "^1.17.0"
opentelemetry-exporter-otlp = "^1.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
